Trekk dto-er ut i egen fil og legg til AvklartKjørtUkeStatus fordi denne trengs av frontend
